externalize interfaces
